Made Platform the super class of Terraform
Moved get_platform to platforms/__init__.py
Moved the functions that get overridden to Platform
Moved cleanup function to Platform
Moved gather_logs to Platform
Moved provision to Platform
